requirements.txt
# Core Data Processing
pyspark==3.5.0
delta-spark==2.4.0

# Workflow Orchestration
apache-airflow==2.7.3
apache-airflow-providers-apache-spark==4.3.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Monitoring & Logging
prometheus-client>=0.18.0

# Utilities
tqdm>=4.65.0
click>=8.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

